pyproject.toml
[project]
name = "bd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "faker>=38.2.0",
    "fastapi>=0.123.9",
    "jinja2>=3.1.6",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
